+++ Weather.Common/Interfaces/ISensorType.cs
@@ -4,7 +4,7 @@ namespace Weather.Common.Interfaces
 {
     public interface ISensorType
     {
-        int SensorTypeId { get; set; }
+        long SensorTypeId { get; set; }
         string Name { get; set; }
         IUnitType UnitType { get; set; }
         bool IsValid { get; }
