+++ Weather.Common/Interfaces/IWeatherRecord.cs
@@ -1,14 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
+using Weather.Common.Entities;
 
 namespace Weather.Common.Interfaces
 {
     public interface IWeatherRecord
     {
-        int WeatherRecordId { get; set; }
+        long WeatherRecordId { get; set; }
         DateTime TimeStamp { get; set; }
         List<ISensorValue> SensorValues { get; set; }
         IWeatherStation WeatherStation { get; set; }
-        int WeatherStationId { get; set; }
+        long WeatherStationId { get; set; }
     }
 }
